FEATURE: use GitHub API for granting badges instead of cloning repos (#6)
* FEATURE: use GitHub API for granting badges instead of cloning repos

Cloning very large repos is an expensive operation that can cause
significant overhead on the server. This change will make the plugin
use the GitHub API for granting badges. The way it works is it keeps
track of commits in the `github_commits` table, and then every 4 hours
it'll check for new commits and grant badges to users who qualify.
The first time the job is run it'll try to build the entire git histroy
in the table. If the GitHub API rate limit us, then the next run will
resume building the history from where it stopped last time.

Providing a GitHub token via site settings is required in order for the
badges to be granted to users. The token doesn't need any permissions
unless one of the target repos is private.

* Add new site setting to enable/disable the plugin

This setting will make the "Settings" button show up for this plugin at
/admin/plugins

* Use the existing github_linkback_access_token setting instead of adding a new one

* Remove freeze