FIX: octokit error with trailing slash in repo name
If the repo url specified in site settings has a trailing slash octokit
will raise an 'invalid_repository' error.

This fix strips the trailing slash from any urls specified in
`github_badges_repos` and adds a model test for this as well as a test
for stripping `.git` from the url.