FIX: Allow different badge repo format and validate setting (#14)
This fix is needed because the logs have had the following error on a few instances:

"Job exception: "$REPO_NAME" is invalid as a repository identifier. Use the repo/user (String) format, or the repository ID (Integer), or a hash containing :repo and :user keys."

This was happening because we were not validating the site setting for github_badges_repos, and allowing the user to put in any value here, so some people were just putting in a repo name instead of a URL.

We now allow repos in the formats:

https://github.com/discourse/discourse.git
https://github.com/discourse/discourse
discourse/discourse

And we also have a proper setting validator class and tests.