FIX: Improve contributors detection (#24)
This plugin ships with 2 types of badges: Contributor and Committer.
Currently, these badges are granted by replicating the git history of the
configured repositories in the database using Github's REST API, and
then looking up users by the emails on the commits and hand out badges.

A commit is considered a 'contribution' if it comes from a pull request
that was not created and merged by the same user. The commits endpoint
of Github's REST API doesn't include whether a commit comes from a PR or
not. So, our only way to detect contributions is to compare the
`committer` and `author` fields of the commit object that we get from
the API. This method worked fairly well up until December 2019 when
Github changed something that made the `committer` always be set to
`GitHub` and not the user who merged the pull request.

So, this commit replaces the REST API with GraphQL which allows us to
include pull request data (if any) with each commit object. This gives
us a reliable way for detecting contributions.